refactor: change course images
